Improving some codes...
