+++ app/src/main/java/com/example/rombeng/view/RombengScreen.kt
@@ -1935,166 +1935,6 @@ fun CategoryItem(text: String, modifier: Modifier = Modifier) {
 //    }
 //}
 
-//Upload Screen
-@Composable
-fun UploadItemScreen(navController: NavController, viewModel: RombengViewModel = viewModel()) {
-
-    // === Buat Fullscreen / Immersive Mode ===
-    val view = LocalView.current
-    LaunchedEffect(Unit) {
-        val window = (view.context as Activity).window
-        window.decorView.systemUiVisibility = (
-                View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
-//                        or View.SYSTEM_UI_FLAG_FULLSCREEN
-                        or View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
-                )
-    }
-
-    var judul by remember { mutableStateOf("") }
-    var harga by remember { mutableStateOf("") }
-    var deskripsi by remember { mutableStateOf("") }
-    var lokasi by remember { mutableStateOf("") }
-    var currentPhotoCount by remember { mutableStateOf(0) }
-    val maxPhoto = 4
-    val remainingQuota = 0
-
-    var showExitDialog by remember { mutableStateOf(false) }
-
-    BackHandler {
-        showExitDialog = true
-    }
-
-    if (showExitDialog) {
-        val activity = LocalActivity.current
-        AlertDialog(
-            onDismissRequest = { showExitDialog = false },
-            title = { Text("Keluar Aplikasi") },
-            text = { Text("Apakah Anda yakin ingin keluar?") },
-            confirmButton = {
-                TextButton(onClick = {
-                    activity?.finish()
-                }) {
-                    Text("Ya")
-                }
-            },
-            dismissButton = {
-                TextButton(onClick = { showExitDialog = false }) {
-                    Text("Tidak")
-                }
-            }
-        )
-    }
-
-    Box(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.White)
-            .systemBarsPadding()
-    ) {
-        // Layout Utama
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .padding(16.dp)
-                .verticalScroll(rememberScrollState()),
-            horizontalAlignment = Alignment.CenterHorizontally
-        ) {
-
-            // Gambar Upload Placeholder
-            Box(
-                modifier = Modifier
-                    .size(100.dp)
-                    .clip(RoundedCornerShape(12.dp))
-                    .background(Color(0xFFEFEFEF))
-                    .clickable {
-                        if (currentPhotoCount <= maxPhoto) currentPhotoCount++
-                    },
-                contentAlignment = Alignment.Center
-            ) {
-                Icon(
-                    imageVector = Icons.Default.Add,
-                    contentDescription = "Tambah Foto",
-                    tint = Color.Gray,
-                    modifier = Modifier.size(36.dp)
-                )
-            }
-
-            Spacer(modifier = Modifier.height(8.dp))
-            Text("Photo: $currentPhotoCount/5")
-
-            Spacer(modifier = Modifier.height(12.dp))
-
-            // Judul
-            OutlinedTextField(
-                value = judul,
-                onValueChange = { judul = it },
-                placeholder = { Text("Judul") },
-                modifier = Modifier.fillMaxWidth(),
-            )
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            // Harga
-            OutlinedTextField(
-                value = harga,
-                onValueChange = { harga = it },
-                placeholder = { Text("Harga") },
-                modifier = Modifier.fillMaxWidth()
-            )
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            // Kategori Dropdown
-            Column(modifier = Modifier.fillMaxWidth()) {
-                KategoriDropdown()
-            }
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            // Deskripsi
-            OutlinedTextField(
-                value = deskripsi,
-                onValueChange = { deskripsi = it },
-                placeholder = { Text("Deskripsi") },
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(100.dp)
-            )
-
-            Spacer(modifier = Modifier.height(8.dp))
-
-            // Lokasi
-            OutlinedTextField(
-                value = lokasi,
-                onValueChange = { lokasi = it },
-                placeholder = { Text("Lokasi") },
-                modifier = Modifier.fillMaxWidth(),
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            // Tombol Unggah
-            UploadButton(
-                remainingQuota = remainingQuota,
-                judul = judul,
-                harga = harga,
-                lokasi = lokasi,
-                onNavigateToPembayaran = {
-                    navController.navigate("pembayaran")
-                }
-            )
-
-            Spacer(modifier = Modifier.height(32.dp))
-        }
-
-
-        // BottomNavBar dipastikan di bawah
-        Box(modifier = Modifier.align(Alignment.BottomCenter)) {
-            BottomNavBar(navController)
-        }
-    }
-
-}
 
 @Composable
 fun KategoriDropdown() {
@@ -2241,164 +2081,13 @@ fun UploadButton(
     }
 }
 
-@Composable
-fun CartScreen(navController: NavController) {
-    Column(
-        modifier = Modifier
-            .fillMaxSize()
-            .background(Color.White)
-    ) {
-        // Header
-        TopAppBar(
-            modifier = Modifier
-                .navigationBarsPadding()
-                .fillMaxWidth(),
-            title = { Text("Keranjang", fontWeight = FontWeight.Bold, color = Color.Black, fontSize = 20.sp) },
-            navigationIcon = {
-                IconButton(onClick = { navController.popBackStack() }) {
-                    Icon(Icons.Default.ArrowBack, contentDescription = "Kembali", tint = Color.Black)
-                }
-            },
-            backgroundColor = Color(0xFFFF8D21),
-            contentColor = Color.Black
-        )
-
-        // Konten Utama
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .weight(1f)
-                .padding(16.dp),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Center
-        ) {
-            Box(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .background(Color(0xFFF3F3F3), RoundedCornerShape(16.dp))
-                    .padding(40.dp),
-                contentAlignment = Alignment.Center
-            ) {
-                Column(horizontalAlignment = Alignment.CenterHorizontally) {
-                    Icon(
-                        imageVector = Icons.Default.ShoppingBag,
-                        contentDescription = "Empty Cart",
-                        modifier = Modifier.size(64.dp),
-                        tint = Color.Black
-                    )
-                    Spacer(modifier = Modifier.height(8.dp))
-                    Text("Tidak ada barang dalam keranjangmu")
-                }
-            }
-        }
-
-        // Informasi Total Pembayaran & Tombol Bayar
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Color(0xFFF9F9F9), RoundedCornerShape(topStart = 24.dp, topEnd = 24.dp))
-                .padding(16.dp)
-        ) {
-            Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
-                Text("Sub Total", color = Color.Gray)
-                Text("Rp.0", color = Color.Gray)
-            }
-            Spacer(modifier = Modifier.height(8.dp))
-            Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
-                Text("Biaya Layanan", color = Color.Gray)
-                Text("Rp.0", color = Color.Gray)
-            }
-            Spacer(modifier = Modifier.height(8.dp))
-            Row(Modifier.fillMaxWidth(), horizontalArrangement = Arrangement.SpaceBetween) {
-                Text("Total Pembayaran", fontWeight = FontWeight.Bold)
-                Text("Rp.0", fontWeight = FontWeight.Bold)
-            }
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            Button(
-                onClick = { /* aksi saat klik bayar */ },
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(48.dp),
-                enabled = false, // karena keranjang kosong
-                shape = RoundedCornerShape(16.dp),
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = Color(0xFFFFD4A3),
-                    contentColor = Color.White,
-                    disabledContainerColor = Color(0xFFFFD4A3),
-                    disabledContentColor = Color.Gray
-                )
-            ) {
-                Text("Bayar")
-            }
-        }
-    }
-}
 
 @Composable
-fun ProfileScreen(navController: NavHostController) {
-    Column(modifier = Modifier.fillMaxSize()) {
-        // Top Bar
-        TopAppBar(
-            title = {
-                Text(
-                    text = "Profil",
-                    fontWeight = FontWeight.Bold,
-                    fontSize = 20.sp,
-                    color = Color.Black
-                )
-            },
-            navigationIcon = {
-                IconButton(onClick = { navController.popBackStack() }) {
-                    Icon(
-                        imageVector = Icons.Default.ArrowBack,
-                        contentDescription = "Back",
-                        tint = Color.Black
-                    )
-                }
-            },
-            backgroundColor = Color.White,
-            contentColor = Color.Black
-        )
+fun UserProfile(){
 
-        // Menu Items
-        Column(
-            modifier = Modifier
-                .fillMaxWidth()
-                .padding(16.dp),
-            verticalArrangement = Arrangement.spacedBy(24.dp)
-        ) {
-            ProfileMenuItem(icon = Icons.Outlined.Person, label = "Profil")
-            ProfileMenuItem(icon = Icons.Outlined.Home, label = "Alamat")
-            ProfileMenuItem(icon = Icons.Outlined.AccountBalanceWallet, label = "Pembayaran")
-            ProfileMenuItem(icon = Icons.Outlined.ChatBubbleOutline, label = "Customer Service")
-            ProfileMenuItem(icon = Icons.Outlined.Settings, label = "Pengaturan")
-        }
-    }
 }
 
-@Composable
-fun ProfileMenuItem(icon: ImageVector, label: String) {
-    Row(
-        verticalAlignment = Alignment.CenterVertically,
-        modifier = Modifier.fillMaxWidth()
-    ) {
-        Icon(
-            imageVector = icon,
-            contentDescription = label,
-            modifier = Modifier
-                .size(40.dp)
-                .padding(end = 16.dp),
-            tint = Color.Black
-        )
-        Text(
-            text = label,
-            fontSize = 16.sp,
-            color = Color.Black
-        )
-    }
-}
+
