+++ app/src/main/java/com/example/rombeng/view/UploadItemScreen.kt
@@ -1,0 +1,202 @@
+package com.example.rombeng.view
+
+import android.app.Activity
+import android.view.View
+import androidx.activity.compose.BackHandler
+import androidx.activity.compose.LocalActivity
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.systemBarsPadding
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Add
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.Icon
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.platform.LocalView
+import androidx.compose.ui.unit.dp
+import androidx.lifecycle.viewmodel.compose.viewModel
+import androidx.navigation.NavController
+import com.example.rombeng.viewmodel.RombengViewModel
+
+@Composable
+fun UploadItemScreen(navController: NavController, viewModel: RombengViewModel = viewModel()) {
+
+    // === Buat Fullscreen / Immersive Mode ===
+    val view = LocalView.current
+    LaunchedEffect(Unit) {
+        val window = (view.context as Activity).window
+        window.decorView.systemUiVisibility = (
+                View.SYSTEM_UI_FLAG_IMMERSIVE_STICKY
+//                        or View.SYSTEM_UI_FLAG_FULLSCREEN
+                        or View.SYSTEM_UI_FLAG_HIDE_NAVIGATION
+                )
+    }
+
+    var judul by remember { mutableStateOf("") }
+    var harga by remember { mutableStateOf("") }
+    var deskripsi by remember { mutableStateOf("") }
+    var lokasi by remember { mutableStateOf("") }
+    var currentPhotoCount by remember { mutableStateOf(0) }
+    val maxPhoto = 4
+    val remainingQuota = 0
+
+    var showExitDialog by remember { mutableStateOf(false) }
+
+    BackHandler {
+        showExitDialog = true
+    }
+
+    if (showExitDialog) {
+        val activity = LocalActivity.current
+        AlertDialog(
+            onDismissRequest = { showExitDialog = false },
+            title = { Text("Keluar Aplikasi") },
+            text = { Text("Apakah Anda yakin ingin keluar?") },
+            confirmButton = {
+                TextButton(onClick = {
+                    activity?.finish()
+                }) {
+                    Text("Ya")
+                }
+            },
+            dismissButton = {
+                TextButton(onClick = { showExitDialog = false }) {
+                    Text("Tidak")
+                }
+            }
+        )
+    }
+
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(Color.White)
+            .systemBarsPadding()
+    ) {
+        // Layout Utama
+        Column(
+            modifier = Modifier
+                .fillMaxSize()
+                .padding(16.dp)
+                .verticalScroll(rememberScrollState()),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+
+            // Gambar Upload Placeholder
+            Box(
+                modifier = Modifier
+                    .size(100.dp)
+                    .clip(RoundedCornerShape(12.dp))
+                    .background(Color(0xFFEFEFEF))
+                    .clickable {
+                        if (currentPhotoCount <= maxPhoto) currentPhotoCount++
+                    },
+                contentAlignment = Alignment.Center
+            ) {
+                Icon(
+                    imageVector = Icons.Default.Add,
+                    contentDescription = "Tambah Foto",
+                    tint = Color.Gray,
+                    modifier = Modifier.size(36.dp)
+                )
+            }
+
+            Spacer(modifier = Modifier.height(8.dp))
+            Text("Photo: $currentPhotoCount/5")
+
+            Spacer(modifier = Modifier.height(12.dp))
+
+            // Judul
+            OutlinedTextField(
+                value = judul,
+                onValueChange = { judul = it },
+                placeholder = { Text("Judul") },
+                modifier = Modifier.fillMaxWidth(),
+            )
+
+            Spacer(modifier = Modifier.height(8.dp))
+
+            // Harga
+            OutlinedTextField(
+                value = harga,
+                onValueChange = { harga = it },
+                placeholder = { Text("Harga") },
+                modifier = Modifier.fillMaxWidth()
+            )
+
+            Spacer(modifier = Modifier.height(8.dp))
+
+            // Kategori Dropdown
+            Column(modifier = Modifier.fillMaxWidth()) {
+                KategoriDropdown()
+            }
+
+            Spacer(modifier = Modifier.height(8.dp))
+
+            // Deskripsi
+            OutlinedTextField(
+                value = deskripsi,
+                onValueChange = { deskripsi = it },
+                placeholder = { Text("Deskripsi") },
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(100.dp)
+            )
+
+            Spacer(modifier = Modifier.height(8.dp))
+
+            // Lokasi
+            OutlinedTextField(
+                value = lokasi,
+                onValueChange = { lokasi = it },
+                placeholder = { Text("Lokasi") },
+                modifier = Modifier.fillMaxWidth(),
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            // Tombol Unggah
+            UploadButton(
+                remainingQuota = remainingQuota,
+                judul = judul,
+                harga = harga,
+                lokasi = lokasi,
+                onNavigateToPembayaran = {
+                    navController.navigate("pembayaran")
+                }
+            )
+
+            Spacer(modifier = Modifier.height(32.dp))
+        }
+
+
+        // BottomNavBar dipastikan di bawah
+        Box(modifier = Modifier.align(Alignment.BottomCenter)) {
+            BottomNavBar(navController)
+        }
+    }
+
+}
